몬스터들에 사용하던 CommonState대신 그냥 int값으로 CommonState 표현하게 변경.
즉, 모든 몬스터의 enum class에는 CommonState의 enum값들이 순서에맞게 먼저 선언되어있긴 해야하지만, 불필요한 클래스생성을 방지 및 코드가 복잡해지는걸 방지.

+++ Source/HyobinsProject/private/Monster/BTS_SetPatrolState.cpp
@@ -19,5 +19,5 @@ void UBTS_SetPatrolState::OnBecomeRelevant(UBehaviorTreeComponent& OwnerComp, ui
 
 	AMonster* const owner = Cast<AMonster>(OwnerComp.GetAIOwner()->GetPawn());
 	checkf(IsValid(owner), TEXT("Monster isn't Valid"));
-	owner->SetCommonState(EMonsterCommonStates::Patrol); 
+	owner->SetCommonState(static_cast<uint8>(EMonsterCommonStates::Patrol)); 
 }
